Breath Life into rrogg-emacs and Measure Starttime
Declare Emacs bankruptcy and start rrogg-emacs!
Set header arguments for the complete file. Tangle all source blocks to
the init.el file in a folder where Emacs will look for it.
This config shall work cross platform on Linux, macOS and Windows
systems in the end. Make sure that ~ is pointing to the users correct
home folder.
Save the time of the start of Emacs in order to monitor how long the
config needs to load. Print that info into the *Messages* buffer.